refactor: GameResultTest에서 하는 테스트 메서드 BlackJackGameTest로 옮김
- command enum 도메인 영역에서 제거로인한 불필요한 테스트 제거
